refactor plant detail into components
